+++ apps/backend/src/modules/currency/core/ports/exchange_rate_repository_port.py
@@ -1,11 +1,10 @@
 from abc import ABC
 from abc import abstractmethod
-from decimal import Decimal
 
 from modules.currency.core.enums.currency_enum import CurrencyEnum
 
 
 class ExchangeRateRepositoryPort(ABC):
     @abstractmethod
-    def get_exchange_rate(self, currency: CurrencyEnum) -> Decimal:
+    def get_exchange_rate(self, currency: CurrencyEnum) -> float:
         pass
